refactor(cli): move logs handler to handlers/logs.rs (500-line gate)
